+++ src/InputMonitors/InputMonitor.cpp
@@ -106,7 +106,7 @@ void InputMonitor::DigitalInterrupt() noexcept
 	}
 }
 
-void InputMonitor::AnalogInterrupt(uint32_t reading) noexcept
+void InputMonitor::AnalogInterrupt(int32_t reading) noexcept
 {
 	const bool newState = reading >= threshold;
 	if (newState != state)
@@ -130,7 +130,7 @@ void InputMonitor::AnalogInterrupt(uint32_t reading) noexcept
 	static_cast<InputMonitor*>(cbp.vp)->DigitalInterrupt();
 }
 
-/*static*/ void InputMonitor::CommonAnalogPortInterrupt(CallbackParameter cbp, uint32_t reading) noexcept
+/*static*/ void InputMonitor::CommonAnalogPortInterrupt(CallbackParameter cbp, int32_t reading) noexcept
 {
 	static_cast<InputMonitor*>(cbp.vp)->AnalogInterrupt(reading);
 }
@@ -255,7 +255,7 @@ void InputMonitor::AnalogInterrupt(uint32_t reading) noexcept
 		break;
 
 	case CanMessageChangeInputMonitorV1::actionChangeThreshold:
-		m->threshold = msg.param;
+		m->threshold = (int32_t)msg.param;
 		rslt = GCodeResult::ok;
 		break;
